+++ src/main/java/com/app/Entities/Skill.java
@@ -5,19 +5,19 @@ package com.app.Entities;
  */
 public class Skill {
 
-    private long id;
+    private int id;
     private String name;
 
-    public Skill(long id, String name) {
+    public Skill(int id, String name) {
         this.id = id;
         this.name = name;
     }
 
-    public long getId() {
+    public int getId() {
         return id;
     }
 
-    public void setId(long id) {
+    public void setId(int id) {
         this.id = id;
     }
